requirements.txt
Pillow==8.5.0
requests==2.26.0
torch==1.11.0
torchvision==0.12.0
git+https://github.com/openai/CLIP.git@main#egg=clip
ultralytics==8.0.14
tk==0.1.0
